Implement authorization using Policy

+++ app/Http/Controllers/QuestionController.php
@@ -8,11 +8,18 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+
+     public function __construct()
+     {
+        $this->middleware('auth', ['except' => ['index','show']]);
+     }
+
     public function index()
     {
         //\DB::enableQueryLog();
@@ -84,10 +91,15 @@ class QuestionController extends Controller
         abort(403, "Access Denied.");
         */
 
+        /*
         if(\Gate::denies('update-question',$question))
         {
             abort(403, "Access Denied.");
         }
+        return view("questions.edit", compact('question'));
+        */
+
+        $this->authorize("update", $question);
         return view("questions.edit", compact('question'));
     }
 
@@ -100,10 +112,14 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        /*
         if(\Gate::denies('update-question',$question))
         {
             abort(403, "Access Denied.");
         }
+        */
+
+        $this->authorize("update", $question);
 
         $question->update($request->only('title','body'));
 
@@ -118,10 +134,14 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        /*
         if(\Gate::denies('delete-question',$question))
         {
             abort(403, "Access Denied.");
         }
+        */
+
+        $this->authorize("delete", $question);
 
         $question->delete();
